requirements.txt
openai>=1.103.0
beautifulsoup4>=4.13.5
google-api-python-client>=2.179.0
langchain-cerebras>=0.5.0
langchain-core>=0.3.75
langchain-google-genai>=2.1.10
langchain-ollama>=0.3.7
langchain-anthropic>=0.3.19
langgraph>=0.6.6
langgraph-checkpoint-sqlite>=2.0.11
lxml>=5.4.0
ollama>=0.5.3
prompt-toolkit>=3.0.52
python-dotenv>=1.1.1
pyyaml>=6.0.3
rapidfuzz>=3.14.0
requests>=2.32.5
rich>=14.1.0
hf_xet>=1.1.10
python-docx>=1.2.0
pymupdf4llm>=0.0.27
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
